requirements.txt
streamlit==1.40.1
extra-streamlit-components==0.1.71
firebase==4.0.1
firebase-admin==6.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
numpy==1.26.4
Pillow==10.4.0
Pyrebase4==4.8.0
setuptools